==> 2. Stock variables_v8.ipynb <==
## 1. Import Relevant Packages

import pandas as pd
import numpy as np
import datetime as dt
import matplotlib.pyplot as plt

import statsmodels.api as sm

from statsmodels.regression.rolling import RollingOLS

from functools import reduce
import math

## 2. Variables created for ANN 

**Market variables** (merges på date)
* Market return (1M)
* Cumulative market return (12-1M)
* Standard deviation (12-1M)

**Stock variables** (merges på date og permno)
* Stock returns (1M) 
* Stock cumulative returns (12-1M) 
* Standard deviation of stock returns (12-1M)
* Alpha(12-1M)
* Beta (12-1M) 
* Idiosyncreatic returns (12-1M)

## 3. Importing and preprocessing data

### 3.1 Importing CRSP data

# OBS: Change file to "LSTM file" when momentum strategy code is discussed 
crsp_m = pd.read_csv('Input.csv', low_memory=False)

# Make Python datetime format
crsp_m['date'] =  pd.to_datetime(crsp_m['date'])

crsp_m=crsp_m.sort_values(by='date')
crsp_m

sorted_crsp = crsp_m.sort_values(['permno','date']).set_index('date')

sorted_crsp.head(11)

### 3.2 Importing Fama French data

# Loading Fama French 3 Factors (only include rows with factors)
ff_3f = pd.read_csv('FF_3.csv', skiprows = 3, nrows=1134, index_col = 0)

# Get in decimal format 
ff_3f = ff_3f / 100

# Checking
ff_3f.head()

# Setting datetime format (in months)
ff_3f.index = pd.to_datetime(ff_3f.index, format= '%Y%m')
ff_3f.index = ff_3f.index.to_period('M')

## 4. Market variables

# Calculating market return
market = pd.DataFrame()
market['Mkt'] = ff_3f['Mkt-RF'] + ff_3f['RF']

# Calculating log and sum of log returns (cumulative market return) over past 11 months
market_log = market.copy()
market_log['Mkt_logret']= np.log(1+market['Mkt'])
market_log['Mkt_cumlog'] = market_log['Mkt_logret'].rolling(11, min_periods=11).sum()

# Get non-log
market['Mkt_cumret'] = np.exp(market_log['Mkt_cumlog'])-1

market['Mkt_std'] = market['Mkt'].rolling(11, min_periods=11).std()

# Lag once in order to ensure that previous months are used for prediction
# Lag once more in order to include skipping period
market_lagged = market.shift(2)

# Include only relevant period
market_lagged = market_lagged[market_lagged.index >= '1928-12']
market_train = market_lagged[market_lagged.index <= '1975-12']
market_test = market_lagged[market_lagged.index > '1975-12']

# Importing MinMax scaler to convert to scale from -1 to 1
from sklearn.preprocessing import StandardScaler

# Defining scaler from -1 to 1
scaler = StandardScaler()

# Fit scaler to training values
scaler = scaler.fit(market_train)

# Scale train data
market_train_scaled = pd.DataFrame(scaler.transform(market_train))

# Scale test data
market_test_scaled =  pd.DataFrame(scaler.transform(market_test))

market_test_scaled.describe()

# Resetting columns and index
market_train_scaled.columns = ['Mkt', 'Mkt_cumret','Mkt_std']
market_train_scaled.index = market_train.index

market_test_scaled.columns = ['Mkt', 'Mkt_cumret','Mkt_std']
market_test_scaled.index = market_test.index

market_train_scaled

**Testing for stationarity**

# Testing for stationarity (i.e. if p-value <= 0.05)
# Import adfuller

from statsmodels.tsa.stattools import adfuller
# Stationarity of market returns
print("Stationarity of market returns:")
result = adfuller(market_lagged['Mkt'])
print('ADF Statistic: %f' % result[0])
print('p-value: %f' % result[1])
print('Critical Values:')
for key, value in result[4].items():
    print('\t%s: %.3f' % (key, value))
    

from statsmodels.tsa.stattools import adfuller
# Stationarity of market cum. returns
print("Stationarity of market cum. returns:")
result = adfuller(market_lagged['Mkt_cumret'])
print('ADF Statistic: %f' % result[0])
print('p-value: %f' % result[1])
print('Critical Values:')
for key, value in result[4].items():
    print('\t%s: %.3f' % (key, value))

# Market file to CSV
market_train_scaled.to_csv('market_train.csv', index=True)
market_test_scaled.to_csv('market_test.csv', index=True)

# Save unscalled market for plotting
market_lagged.to_csv('market_unscalled.csv', index=True)

## 3. Stock variables 

### 3.1 Return

# Make into pivot format
sorted_crsp_p = sorted_crsp.pivot(columns='permno', values=['ret'])

sorted_ret = sorted_crsp_p['ret']

# Ensure that there is no N/A values in between returns
sorted_ret[sorted_ret.bfill().notnull() & sorted_ret.ffill().notnull()] = sorted_ret.fillna(0)
sorted_ret[sorted_ret.bfill().notnull() & sorted_ret.ffill().notnull()] = sorted_ret.fillna(0)

# Return CSV file
m_sorted_ret = sorted_ret.copy()

sorted_ret.index = sorted_ret.index.to_period('M')
sorted_ret = sorted_ret[sorted_ret.index >= '1927-01']
sorted_ret_lagged = sorted_ret.shift(1)

sorted_ret_lagged.head()

# To CSV
sorted_ret_lagged.to_csv('input_ret.csv', index=True)

### 3.2 Beta

# Defining beginning period for FF
ff_3f = ff_3f[ff_3f.index >= '1927-01']

# Calculating excess return (output variable for CAPM regression)
excess_return = pd.DataFrame()
for stock in sorted_ret.columns:
    excess_return[stock] = sorted_ret[stock] - ff_3f['RF']

excess_return

# Estimating beta
CAPM_dict = {}

for stock in excess_return.columns:
    y=excess_return[stock]
    X=sm.add_constant(ff_3f['Mkt-RF'])
    rols = RollingOLS(y, X, window=11, min_nobs=11) #don't include NaN values
    rres = rols.fit()
    
    CAPM_dict[stock] = rres.params

# Creating dataframe for beta
beta = pd.DataFrame()
for stock in excess_return.columns:
    beta[stock] = CAPM_dict[stock]['Mkt-RF']

#Lag beta once to get one month skipping period
beta_lagged = beta.shift(1)

# Save as csv file
beta_lagged.to_csv('input_beta.csv', index=True)

### 3.3 Idiosyncratic Momentum (Error Term)

# Estimating alpha and beta using Fama/French three-factor model

CAPM_dict = {}

for stock in excess_return.columns:
    y=excess_return[stock]
    X=sm.add_constant(ff_3f[['Mkt-RF', 'SMB', 'HML']])
    rols = RollingOLS(y, X, window=11, min_nobs=11) #Alternative set window to 36 and min_nobs to 12
    rres = rols.fit()
    
    CAPM_dict[stock] = rres.params

# Calculating and creating DataFrame for idiosyncratic mom / error term (OBS Blitz uses 36 month - discuss)
error_df = pd.DataFrame()
for stock in excess_return.columns:
    error_df[stock] = excess_return[stock] - CAPM_dict[stock]['const']-ff_3f['Mkt-RF']*CAPM_dict[stock]['Mkt-RF']-ff_3f['SMB']*CAPM_dict[stock]['SMB']-ff_3f['HML']*CAPM_dict[stock]['HML']    

# Lag error once to get one-month skipping period
error_lagged = error_df.shift(1)

# Save as csv file
error_lagged.to_csv('input_idio.csv', index=True)

### 3.4 Alpha

# Creating alpha dataframe
alpha_df = pd.DataFrame()
for stock in excess_return.columns:
    alpha_df[stock]=CAPM_dict[stock]['const']

# Lag alpha once to get one-month skipping period
alpha_lagged = alpha_df.shift(1)

# Save as csv file
alpha_lagged.to_csv('input_alpha.csv', index=True)

### 3.5 Standard deviation

# Calculating 12-1M rolling stddev
std_df = sorted_ret.rolling(11, min_periods=11).std()
std_df.head(11)

# Lag std once to get one month skipping period
std_lagged = std_df.shift(1)

# Standard deviation to CSV
std_lagged.to_csv('input_std.csv', index=True)

### 3.6 Cumulative return

# Calculating log return
log_df = np.log(1+sorted_ret)
log_df

# Calculating cumulative return
cumret_log = log_df.rolling(11, min_periods=11).sum()
cumret = np.exp(cumret_log)-1

# Lag cum.return once to get one-month skipping period
cumret_lagged = cumret.shift(1)

# Cumulative return CSV file
cumret_lagged.to_csv('input_ret_cum.csv', index=True)

## 4. Output variable

### 4a. Regression output variable 

# Preprocessing output variable
y_file = pd.read_csv('holding_df.csv', low_memory=False)
y_file = y_file[['permno', 'date', 'ret']]
y_file['date'] =  pd.to_datetime(y_file['date'])
y_reg = y_file.set_index('date')

# Resample to month
y_reg.index = y_reg.index.to_period('M')
y_reg = y_reg[y_reg.index > '1930-12']

# Check the number of permnos
y_reg.nunique()

y_reg

# Sort values
y_reg = y_reg.sort_values(['date','permno'])

y_reg.drop('permno', axis=1, inplace=True)

y_reg

# Merge delisting return into missing_dlret dataframe
y_reg = pd.merge(y_reg, ff_3f['RF'], left_on=y_reg.index, right_on=ff_3f.index, how='left')

y_reg.rename(columns={'key_0': 'date'},inplace=True)

y_reg = y_reg.set_index('date')

y_reg['excess_returns'] = y_reg['ret'] - y_reg['RF']

y_reg.drop(['ret','RF'], axis=1, inplace=True)

y_reg

#y_reg = y_reg.ret * 100

# Output (target) CSV file
y_reg.to_csv('y_reg.csv', index=True)
