==> 4a.-Machine-learning-models_vRF.ipynb <==
import pandas as pd
import numpy as np
import pickle

#url = 'https://raw.githubusercontent.com/MariaRosendal/Enhancing-Price-Momentum-with-RNN/main/'
url= ""

# 1. Set `PYTHONHASHSEED` environment variable at a fixed value
#seed_value= 123
#import os
#os.environ['PYTHONHASHSEED']=str(seed_value)

# 2. Set `python` built-in pseudo-random generator at a fixed value
#import random
#random.seed(seed_value)

# 3. Set `numpy` pseudo-random generator at a fixed value
#import numpy as np
#np.random.seed(seed_value)

# 4. Set `tensorflow` pseudo-random generator at a fixed value
#import tensorflow as tf
#tf.random.set_seed(seed_value)

# 5. For layers that introduce randomness like dropout, make sure to set seed values 
#model.add(Dropout(0.25, seed=seed_value))
#6 Configure a new global `tensorflow` session
#from keras import backend as K
#session_conf = tf.ConfigProto(intra_op_parallelism_threads=1, inter_op_parallelism_threads=1)
#sess = tf.Session(graph=tf.get_default_graph(), config=session_conf)
#K.set_session(sess)

## 1. Importing input and output variables

# Import X 
X_train = pickle.load(open('X_train_scaled.pkl', 'rb' ))
X_test = pickle.load(open('X_test_scaled.pkl', 'rb' ))

# Import X for MLP and RF 
#X_train_mlp = pickle.load(open('X_train_vec.pkl', 'rb' ))
#X_test_mlp = pickle.load(open('X_test_vec.pkl', 'rb' ))

# Checking for NaN values
print(np.count_nonzero(np.isnan(X_train)), np.count_nonzero(np.isnan(X_test)))

X_train.shape, X_test.shape, 

# Importing y
y = pd.read_csv(url+'y_reg.csv')
y['date'] =  pd.to_datetime(y['date'])
y = y.set_index(['date'])

# Defining length of y_train and y_test
y_train = y[y.index <= '1975-12-01']
y_test = y[y.index > '1975-12-01']

y_train.shape, y_test.shape

# Reformating y
y_train = y_train.values.ravel()
y_test = y_test.values.ravel()
y_train.shape, y_test.shape

y_train

## 2. Models

# Importing packages
import tensorflow as tf
import matplotlib.pyplot as plt

import math
import keras
from keras.models import Sequential
from keras.layers import LSTM
from keras.layers import BatchNormalization 
from keras.layers.core import Dense, Dropout
from keras.wrappers.scikit_learn import KerasRegressor
from keras.callbacks import EarlyStopping

from sklearn.metrics import mean_squared_error

from sklearn.model_selection import TimeSeriesSplit, GridSearchCV
from sklearn.metrics import r2_score

### Linear Regression

# Reshaping X to vectors
n_input = X_train.shape[1] * X_train.shape[2]
X_train_mlp = X_train.reshape(len(X_train),n_input)
X_test_mlp = X_test.reshape(len(X_test),n_input)

from sklearn.linear_model import LinearRegression

m_lin = LinearRegression()
m_lin.fit(X_train_mlp, y_train)

lin_predict = m_lin.predict(X_test_mlp)

testScore = math.sqrt(mean_squared_error(y_test, lin_predict))
print('Test Score: %.2f RMSE' % (testScore))
print(testScore)
print((r2_score(y_test, lin_predict)))

lin = pd.DataFrame(lin_predict)
lin.nunique()

pickle.dump(lin_predict, open('lin_predict.pkl', 'wb' ))

### Random Forest

X_train_mlp.shape,  X_test_mlp.shape

from skopt.space import Real, Integer
from skopt import BayesSearchCV

# Ensuring data keeps temporal element when split
tscv = TimeSeriesSplit(n_splits=9) 

from sklearn.ensemble import RandomForestRegressor
import sklearn

m_rf = RandomForestRegressor()

n_estimators = 300

#parameters_rf = {'n_estimators':[50, 100, 200, 300], 'max_depth':[3, 5, 7], 'min_samples_split':[3, 5, 10, 50]}
param_opt = dict(max_depth=Integer(1,10), max_features=Integer(3,108))

bayesian = BayesSearchCV(
     m_rf,
     param_opt,
     n_iter=32,
     random_state=100,
     cv=tscv,
     verbose = 0, n_jobs=1
 )

bayesian_result = bayesian.fit(X_train_mlp, y_train)

# Best model parameters
print('According to BayesianOptimization the best parameters are : ')
print('max_depth : ' + str(bayesian_result.best_params_['max_depth']))
print('max_features : ' + str(bayesian_result.best_params_['max_features']))

# Selecting best model
print("Best: %f using %s" % (bayesian_result.best_score_, bayesian_result.best_params_))
means = bayesian_result.cv_results_['mean_test_score']
stds = bayesian_result.cv_results_['std_test_score']
params = bayesian_result.cv_results_['params']
for mean, stdev, param in zip(means, stds, params):
    print("%f (%f) with: %r" % (mean, stdev, param))

# Defining best model parameters
max_depth = bayesian_result.best_params_['max_depth']
max_features = bayesian_result.best_params_['max_features']

#best_rf = RandomForestRegressor(n_estimators=n_estimators, max_depth=max_depth, max_features=max_features)
best_rf = RandomForestRegressor(n_estimators=n_estimators, max_depth=max_depth, max_features=max_depth)

best_rf.fit(X_train_mlp, y_train)

print("Random forest:\t\t\t" + str(best_rf.score(X_test_mlp, y_test)))

rf_predict = best_rf.predict(X_test_mlp)

print(r2_score(y_test, rf_predict))

r2_df = pd.DataFrame()
r2_df['y'] = y_test
r2_df['y_hat'] = rf_predict
r2_df['difference'] = (r2_df['y']-r2_df['y_hat'])

r2_df = pd.DataFrame()
r2_df['y'] = y_test
r2_df['y_hat'] = rf_predict
r2_df['difference'] = (r2_df['y']-r2_df['y_hat'])
r2_df['y_2'] = r2_df['y']**2
r2_df['dif2'] = r2_df['difference']**2
y_2 = r2_df['y_2'].sum() 
dif2 = r2_df['dif2'].sum()
print(1 - (dif2 / y_2))

rf = pd.DataFrame(rf_predict)
rf.nunique()

pickle.dump(rf_predict, open('rf_predict.pkl', 'wb' ))
