==> 5a. MOM strategy_preparing_result_file.ipynb <==
## 1. Import Relevant Packages

import pandas as pd
import numpy as np
from pandas.tseries.offsets import *
from scipy import stats
import datetime as dt

import matplotlib.pyplot as plt
import statsmodels.api as sm

## 2. Daily returns

### Load daily returns

# Loading daily dataframe
d_ret = pd.read_csv('dailyreturns_1931.csv', low_memory=False)
#d_ret = pd.read_csv('dailyreturns.csv', low_memory=False)

start_date = '1931-01-01'

# Defining result starting data
d_ret = d_ret[d_ret['date'] >= start_date]

# Make all column names lower case
d_ret.columns = d_ret.columns.str.lower()

# Dropping irrelevant columns
d_ret.drop(['shrcd', 'exchcd'], inplace = True, axis=1)

d_ret = d_ret.rename(columns={'date':'daily_date'})

# Making date column to datetime
d_ret1 = d_ret.copy()
d_ret1['daily_date'] = pd.to_datetime(d_ret1.daily_date)

# Make new column for merging with monthly rank
d_ret1['date_merge'] = d_ret1['daily_date']
d_ret1['date_merge'] = d_ret1['date_merge'].dt.strftime("%Y/%m")

d_ret1.groupby('daily_date').head()

d_ret.sort_values('daily_date')

# Save as CSV
d_ret1.to_csv('daily_1931.csv', index=False)

d_ret1[d_ret1['permno'] == 91103].tail(30)

### Load holding period DataFrame

# Taking the relevant columns from the holding df 
h_to_daily = pd.read_csv('holding_df.csv', low_memory=False)
h_month = h_to_daily.copy()
h_to_daily = h_to_daily[['permno', 'hdate2', 'momr']]
h_to_daily.rename(columns={'hdate2':'monthly_date'}, inplace=True)

h_month

# To Python datatime fomat
h_to_daily['monthly_date'] = pd.to_datetime(h_to_daily.monthly_date)

# Create column with date for merging 
h_to_daily['date_merge'] = h_to_daily['monthly_date']
h_to_daily['date_merge'] = h_to_daily['date_merge'].dt.strftime("%Y/%m")

h_to_daily.groupby()

# Merge on date and permnos to get daily returns
daily = pd.merge(d_ret1, h_to_daily, on=['date_merge', 'permno'], how='left')

daily['ret'] = pd.to_numeric(daily['ret'], errors='coerce') 

daily.head()

daily

# Exclude when they are not in universe
daily_filtered = daily[daily['momr'].notna()]

daily_filtered.groupby(['daily_date']).head()

# Ensure ret is float value
daily_filtered['ret'] = daily_filtered['ret'].astype(float)

# Create datafame for portfolio return calculation
port_d_return = pd.DataFrame(daily_filtered[['daily_date', 'siccd', 'ret', 'momr']])

# Calculating mean (portfolio returns) for a given date, momentum ranking
port_d_return = port_d_return.groupby(['daily_date','momr'])['ret'].mean().reset_index()
port_d_return_t = port_d_return.copy()

port_d_return_t

# Check mean daily return
port_d_return_t.groupby('momr')['ret'].describe()[['count','mean','std']].reset_index()

# Transpose portfolio layout to have columns as portfolio returns
port_d_return_t = port_d_return.pivot(index='daily_date', columns='momr', values='ret')

# Add prefix port in front of each column
port_d_return_t = port_d_return_t.add_prefix('P')

port_d_return_t.head()

# Creating long-short portfolio
port_d_return_t = port_d_return_t.rename(columns={'P1.0':'losers', 'P10.0':'winners'})
port_d_return_t['long_short'] = port_d_return_t.winners - port_d_return_t.losers

port_d_return_t

port_d_return_t.describe()

# Ensuring no NAN-values
#port_d_return_t['long_short'].isnull().sum()
port_d_return_t[port_d_return_t['long_short'].isnull()]

# Dropping NA values
port_d_return_t = port_d_return_t[port_d_return_t.index != '1985-09-27']
port_d_return_t = port_d_return_t[port_d_return_t.index != '2012-10-29']
port_d_return_t = port_d_return_t[port_d_return_t.index != '2017-09-30']

# Ensure no N/A
port_d_return_t['long_short'].isnull().sum()

port_d_return_t['losers'].isnull().sum()

port_d_return_t['winners'].isnull().sum()

# Defining long_short return
long_short_ret = port_d_return_t['long_short']
winners_ret = port_d_return_t['winners']
losers_ret = port_d_return_t['losers']

# Save as CSV
long_short_ret.to_csv('MOM_1931.csv', index=True)
winners_ret.to_csv('Winners_1931.csv', index=True)
losers_ret.to_csv('Losers_1931.csv', index=True)

### Daily market data

# Loading Fama French 3 Factors (only include rows with factors)
ff = pd.read_csv('FF3Fdaily.csv', skiprows = 3, nrows=24896, index_col = 0)

# Get in decimal format 
ff = ff / 100

# Setting datetime format
ff.index = pd.to_datetime(ff.index, format = '%Y%m%d')

# Results starting dates
ff = ff[ff.index >= start_date]

ff

# Testing if there is any errors 
ff['RF'] = pd.to_numeric(ff['RF'], errors='coerce') 
ff['Mkt-RF'] = pd.to_numeric(ff['Mkt-RF'], errors='coerce') 

ff['Mkt-RF'].isnull().sum()

# Calculating market return
ff['Mkt'] = ff['Mkt-RF'] + ff['RF']
market = ff['Mkt']

# Setting index to UTC (for it to work with Pyfolio)
market.index = pd.to_datetime(market.index).tz_localize('UTC')
market.index.name = 'date'

market

# To as CSV
market.to_csv('market_benchmark_1931.csv', index=True)

# Defining RF
risk_free = ff['RF']

# Setting index to UTC (for it to work with Pyfolio)
risk_free.index = pd.to_datetime(risk_free.index).tz_localize('UTC')
risk_free.index.name = 'date'

# To as CSV
risk_free.to_csv('risk_free_1931.csv', index=True)
