==> 5d. LSTM_results.ipynb <==
import pandas as pd
import numpy as np
from pandas.tseries.offsets import *

from scipy import stats
import datetime as dt

import matplotlib.pyplot as plt
import statsmodels.api as sm

import pickle

# Loading stocks
y_pred1 = pickle.load(open('lstm_predict.pkl', 'rb' ))
y_pred2 = pickle.load(open('lstm_predict_2.pkl', 'rb' ))
y_pred3 = pickle.load(open('lstm_predict_3.pkl', 'rb' ))
y_pred4 = pickle.load(open('lstm_predict_4.pkl', 'rb' ))
y_pred5 = pickle.load(open('lstm_predict_5.pkl', 'rb' ))

y_pred1.shape, y_pred2.shape, y_pred3.shape, y_pred4.shape, y_pred5.shape

y_pred = np.mean(np.array([y_pred1, y_pred2, y_pred3, y_pred4,y_pred5]), axis=0 )

y_pred

# Importing y
y = pd.read_csv('y_reg.csv')
y['date'] =  pd.to_datetime(y['date'])
y = y.set_index(['date'])
y_test = y[y.index > '1975-12-01']
y_test = y_test.values.ravel()

# Testing R2
r2_df = pd.DataFrame()
r2_df['y'] = y_test
r2_df['y_hat'] = y_pred
r2_df['difference'] = (r2_df['y']-r2_df['y_hat'])
r2_df['y_2'] = r2_df['y']**2
r2_df['dif2'] = r2_df['difference']**2
y_2 = r2_df['y_2'].sum() 
dif2 = r2_df['dif2'].sum()
print(1 - (dif2 / y_2))

spear = []
for i in range(len(y_pred)//500):       
    spear.append(stats.spearmanr(y_pred[i*500:(i+1)*500], y_test[i*500:(i+1)*500])[0])

np.mean(spear),np.std(spear,ddof=1)

stats.ttest_1samp(spear, 0.0)

# Loading stock universe
universe_test = pd.read_csv('universe_test.csv')

# Insert predictions
universe_test.insert(1, "y_hat", y_pred)

y_rank = universe_test.copy()

y_rank.info()

y_rank

y_rank['y_hat'].nunique()

y_rank.shape

# For each date, ranking stocks into 1-10 portfolios based on cumret
# 1 = loser, 10 = winner portfolio 

y_rank = y_rank.groupby('date').apply(lambda x: x.sort_values(['y_hat'], ascending = False)).reset_index(drop=True)
y_rank['rank'] = y_rank.groupby('date')['y_hat'].transform(lambda x: pd.qcut(x, 10, labels=False))

# For visualization: Shift momr from 0-9 to 1-10 
y_rank['rank']=1+y_rank['rank']

# Average cumulative return for each portfolio
y_rank.groupby('rank')['y_hat'].mean()

### 2. Merging daily data with LSTM rank

# Loading daily returns
d_ret1 = pd.read_csv('daily_1931.csv')
d_ret1['daily_date'] = pd.to_datetime(d_ret1.daily_date)

d_ret1

# Preparing date colmn in "universe" for merge

#Rename date column to merge
y_rank.rename(columns={'date':'monthly_date'}, inplace=True)

#Make datatime format
y_rank['monthly_date'] = pd.to_datetime(y_rank.monthly_date)

# Create date for merging
y_rank['date_merge'] = y_rank['monthly_date']
y_rank['date_merge'] = y_rank['date_merge'].dt.strftime("%Y/%m")

# Merge on monthly date
daily = pd.merge(d_ret1, y_rank, on=['date_merge', 'permno'], how='left')

# Looking at daily data
daily.head()

# Filter on non-nans
daily_filtered = daily[daily['rank'].notna()]

# Ensure returns are number format
daily_filtered['ret'] = daily_filtered['ret'].astype(float)

# Creating dateframe for portfolio calculation
port_d_return = pd.DataFrame(daily_filtered[['daily_date', 'ret', 'rank']])

# Calculating mean (portfolio return), for a given date, rank
port_d_return = port_d_return.groupby(['daily_date','rank'])['ret'].mean().reset_index()
port_d_return_t = port_d_return.copy()

# Check portfolio return
port_d_return_t.groupby('rank')['ret'].describe()[['count','mean','std']].reset_index()

# Transpose portfolio layout to have columns as portfolio returns
port_d_return_t = port_d_return.pivot(index='daily_date', columns='rank', values='ret')

# Add prefix port in front of each column
port_d_return_t = port_d_return_t.add_prefix('P')

port_d_return_t.head()

# Creating long-short portfolio
port_d_return_t = port_d_return_t.rename(columns={'P1.0':'losers', 'P10.0':'winners'})
port_d_return_t['long_short'] = port_d_return_t.winners - port_d_return_t.losers

port_d_return_t

# Ensuring no NAN-values
port_d_return_t['long_short'].isnull().sum()

# OBS FILLING WITH NA - should not be necessary with new file
port_d_return_t['long_short'] = port_d_return_t['long_short'].fillna(0)
port_d_return_t['long_short'].isnull().sum()

# Defining long_short return
long_short_ret = port_d_return_t['long_short']
long_short_ret

# Dropping NA values
long_short_ret = long_short_ret[long_short_ret.index != '1985-09-27']
long_short_ret = long_short_ret[long_short_ret.index != '2012-10-29']
long_short_ret = long_short_ret[long_short_ret.index != '2017-09-30']

# Save to CSV
long_short_ret.to_csv('lstm_results.csv', index=True)

long_short_ret
